DB changes for resource locking
This patch adds engine_id and atomic_key column to the resource
table for locking the resource and making all operations atomic.
Also added the db api to update the resource table.

The engine which updates the db first, gets the lock on the
resource. Only updates with this engine are possible. Updates with
other engine will not go through as they will fail to find the
matching record in the db.

Change-Id: Id8c3df7a439347652d3feca7481cd72762bf30ea
Implements: blueprint convergence-resource-locking