Restart haproxy in AWS LB during boot
One of possibilities of AWS LoadBalancer usage is to pass it the list of
instance id's to manage already in the template during create.
This patch actually enables this scenario by (re)starting the haproxy
service to catch the created config with initial servers in it.

Note that the other usage scenario (passing the LB id to the *Group
resource) was not affected as in this case there is implicit update of
AWS LB resource that restarts the haproxy service.

Appropriate functional/scenario test is coming in a subsequent patch.

Change-Id: I85a6d67384e2174cba6e1e7e9666b854e11ab5d3