Use configurable flavor for AWS LoadBalancer
Add another "magic" parameter 'LbFlavor' that will be passed inside the
template for AWS LoadBalancer resource, just like 'KeyName' we have
already.

It is a pity that default OpenStack flavors have such a big gap between
m1.tiny (too small for most standard cloud images) and m1.small
(overkill most of the time). This patch enables choice for the flavor of
the AWSLB instance, which should help decreasing the load for testing this
resource in our integration tests as m1.heat is perfectly enough,
and keeps the previous behaviour as default.

Change-Id: Ib0544493a495baf056c12f8facc769f91d438114