Patch validate methods with mock instead of mox
There are huge numbers of tests where we patch validate methods for
constraints. Currently we use for it mox, but it requires us to add
ReplayAll and VerifyAll in tests, where we use mock. So let's teach it
to use mock, it should make easier migration on mock in the future.

Change-Id: Ia82b3507b8699ffa0db20bd3c24279263ff9cc4b