Cancel nested stack updates when appropriate
If a stack update is cancelled while the update of a nested stack is being
processed, the nested stack update is no longer stopped since the
implementation changed in Kilo to use the RPC API to control nested stacks.

To alleviate this, use the RPC API to cancel the (nested) stack update when
the resource update is cancelled. This results in a rollback of the nested
stack, so if a subsequent attempt to roll back the ensted stack by doing a
stack update fails due to the rollback already being in progress, just
ignore the error and wait for the existing operation to complete.

Change-Id: I3eacf216194517c995732a3f4de876551b70fe91
Closes-Bug: #1446252