Do not verify requirements on extension loading
The heat gate is broken due to other libraries pinning
on pbr<1.0

Stevedore extension loading currently has verify_requirements=True.
This change removes this to use the default of False. Loading extensions
is likely the wrong place to be enforcing requirements versions of
libraries that heat depends on, so lets not do it.

This is likely to also improve heat robustness for deployments with
inconsistent dependencies, not just this current heat breakage.

Change-Id: Ifee37e00412e824eead9ba1435ea5a328ff3df58
Closes-Bug: #1458399
(cherry-picked from 4894441ebcea56fa9ec8479a52510f6d7c49ec80)