ASG scaling account for cooldown timestamp & in-progress
There are cases where it takes a long time to create a new resource as
requested by the scaling operation on an ASG resource, for instance, a
nova server creation followed by a complex SoftwareDeployment. During
this process, additional alarms may come in but failed to be blocked
by the current cooldown checking mechanism because the very first
timestamp has yet to be generated. This is leading to unexpected size
adjustment to the ASG.

This patch augments the existing cooldown checking mechanism with a
scaling-in-progress test so that additional alarms arriving during
the very first scaling operation will be ignored.

Closes-Bug: #1375156

Conflicts:
	heat/engine/resources/openstack/heat/scaling_policy.py
	heat/scaling/cooldown.py
	heat/tests/autoscaling/test_heat_scaling_group.py
	heat/tests/autoscaling/test_heat_scaling_policy.py
	heat/tests/autoscaling/test_scaling_group.py
	heat/tests/autoscaling/test_scaling_policy.py

(cherry-picked from b76881b8bcb3e381c5b993a0c84afd37abd8ea5b)
Change-Id: Ib8aa83eed366df7097c9cbb9247eca866ae4b620