Conditionally expose resources based on available services
If the required keystone service for a given resource does
not exist in the keystone catalog, raises
StackResourceUnavailable.

It will be effective only if default_client_name is defined
in the given resource.

It also fixes other test cases, where is_service_available
needs to be returning True

implements blueprint keystone-based-resource-availability
Closes-bug: #1388047

Change-Id: I92afa9ffc3a3333b46dc25921cf7f982777cba76