Refactor Adding service_types to client_plugins
As part of the review https://review.openstack.org/#/c/192529
service_types was added to client_plugins to support the
conditional resource plugin availability.

I found its good to organize this code better by adding
str_key for each key defined in the service_types
so that we no longer have to depend on the index when
multiple service_types are available for a client plugin.

implements blueprint: keystone-based-resource-availability

Change-Id: I352525cce991f664abb28df9569d498cd44baf18