Specify network for server in test_validation
This test does not wait, when all resources will be created. However,
it does not mean, that these resources will not be in create in 
progress state. The error, mentioned in the bug, happens, when nova 
tries to create server and see two available networks (private and 
heat-net). Current patch prevents error by adding network
property to Server resource.

Change-Id: I7c1f4931976ae3a899563b1a9da9f8bc7564806c
Closes-Bug: #1467529