Support time-delays in server.resize
Resize operation in nova(that we are for flavor update) is
asynchronous. So the server may be ACTIVE when we are checking
server-resize operations(not RESIZE as was excepted).
We need to wait for a while till the server becomes RESIZE.

The patch add support of this case to Server resource in heat.
Closes-bug: #1471135

Change-Id: I19914081996104c5f2144a6de795173f26774c21