support list resources with details
Currently, users can list resources without details, or show
specific resource with details. However, they cannot list all
resources with details. It is desirable to support that.
With consistency concern, we should use {list URI}/detail as
pattern for resources detail list just like other detail list
API, but this pattern will conflict with
{resources list URI}/{resource name} when resoruce name
is 'detail', so implement this feature as a param in list API
seems to be this only option in this case.

DocImpact
APIImpact

Change-Id: I59f464d1803009d7ae2ec35d9fdf0671096967e4
Closes-Bug: #1467268