Adopt @six.python_2_unicode_compatible decorator
On Python 3, the decorator does nothing. On Python 2, it aliases the
__str__ method to __unicode__ and creates a new __str__ method that
returns the result of __unicode__() encoded with UTF-8.

All instances of __unicode__ have been removed with the
aforementioned decorator in place instead.

Note: There are some instances of __str__ which aren't handled but
will be once we have actual python34 tests running in the gate.

partial blueprint heat-python34-support

Change-Id: I7271a2581e1c2bbc282933c7da73db810c7e09db