Refactor and correct purge_deleted
This method had not been updated in a while
to account for changes in the db schema. Also,
it was refactored to use bulk delete statements
rather than iteration as the iteration approach
ran quite a long time and consumed quite a bit
of resources.

Change-Id: Ia49e801d3c244a1fabb8441e7eaa22b88e5fcb56
Closes-Bug: #1439518
(cherry picked from commit d3def2acee7e836263319bd64471816cf1de2540)