Return EIP resource name during template validation
In template if a resource tries to reference an AWS EIP resource,
stack create will fail.
This is because during the stack validation phase IP is not generated
yet, and an empty string is returned by the intrinsic function
Ref.

Like all other resources we need to return the resource name if the
physical resource is not created yet.

Change-Id: I310af771ec414f8f5cda3de55861fa960673e7ff
Closes-Bug: #1473984