Add functional test for resource exposure
These simple tests check that Sahara resources can not be created
and are not listed in the resource-type-list.

If we ever decide to install Sahara on the Heat functional tests gate,
tests should be changed to use other not installed but in principle
supported service and resources, as these tests will be skipped in this
case.

Related blueprint keystone-based-resource-availability

Change-Id: I6e692587f49b6c34c5e99b8d26bb6e60b7ce7af5