Convergence: Fix concurrent update
Consider a stack-create which is still in-progress with some of the
resources in INIT COMPLETE. Now when a update is triggered and these
resources also exist in new template, it will be physically created
by this update traversal, however the template id is not updated.
As a result, these resources will be considered for clean up and
deleted, which should not be the case.

This patch ensures that even when creating the resource, the
particular worker will update its present template id.

Change-Id: I42836baaa9ed0ecc188d633c1f8492f73068874c
Closes-Bug: #1476626