Remove patch_previous_parameters() from Environment
As noted in https://review.openstack.org/#/c/154618/ reviews,
the Environment class already has a good mechanism to enable
initializing with an existing environment or parameters, and
then subsequently load additional parameters.

Switching to this method makes things cleaner when we move to
a full PATCH update, e.g one which enables merging all of a
new environment with the existing one, not just parameters,
so refactoring to remove patch_previous_parameters enables
this to be handled more consistently.

Tests have been removed for the removed code, but note that
the updated logic in service.py is covered by existing tests
in test_engine_service which prove the previous parameter
patching behavior is maintained.

Change-Id: Idb2c00d0486b80a8aabf7902b8a7df2e13f9e914