Reuse existing environment on PATCH update
Extend the "existing" option to the engine update so we reuse
all of the environment, not only the parameters, which means
that when -x is specified during a stack-update we'll maintain
any existing resource_registry (and files), then merge any
environment passed during the update into those existing.

This allows the users to optionally omit the environment when
performing a stack update, which is particularly convenient
if you're updating a stack which has a large or complex
environment.

Co-Authored-By: Ton Ngo <ton@us.ibm.com>
Partial-Bug: 1224828
Change-Id: I3033063a06c3913a1ae89ea1ec1fdac688ebd2cf