Fix/refactor test_stack_update_existing_parameters
This test is supposed to prove that passing the "existing" flag to
an update_stack call merges existing parameters with any new ones
passed into the update, but atm it strangely starts with a stack
containing zero existing parameters, and uses some convoluted mocks
to prove that the new parameters are used.  So AFAICT it's not
really testing what is required and it's pretty hard to read.

Refactoring so we look at the environment passed into the Stack
constructor (mocked with mock) enables us to compare the before/
after parameters and ensure the existing and new ones are correctly
combined as expected.

Change-Id: I1178a5127d9fd94fc38cf6d80a88ca75054f2d57