ReST API: Refactor error handler as a context manager
The format parse method was making nasty use of magic strings that were
intended only as error messages to the user in order to switch between two
different behaviours that are each exercised in exactly one place. Since it
is basically just a method for consistent error handling, refactor it as a
context manager.

Change-Id: Ibc33cc5186601a2669652484b2360c8123d263cd