resource failure causes nested stacks to be rolled back
an update_cancel RPC call for a nested stacks whenever an update
operation was cancelled in the parent stack.Unfortunately this always
triggers a rollback.
In this change, resource failure with nested stacks will cancel any
in-progress update but not roll back by default (roll back only when
rollblack cancel exception raised).
The Operator can retry in this way without waiting for timeout
or overall roll back occurred.
Change-Id: I94d75a21367f39c17a9b40b5d23405a28873cd1a
Closes-Bug: #1475057