OS::Cinder::VolumeType help doc string contains redundant info
All resource types except the volumetype is provided without example template.
But volumeType is provided with example, which is not consistent with other
resource types and it introduce maintenance overheat.
This patch remove the example template provided in the doc string
of heat/engine/resources/openstack/cinder/cinder_volume_type.py

Change-Id: Ic7f2a2f2a289f52486551770187cf3def8a4b942
Closes-Bug: #1481312