Clarify name of scaling resource_templates() function
The naming in this function dates from an earlier era, before we had the
ResourceDefinition class. This modifies the function name from
resource_templates() to member_definitions() and also modifies the
parameter/variable names to make their meanings clearer. The documentation
is also improved with a more thorough description of what the function
does.

Change-Id: Ie25a8de380ffc4a6eb4d5934118452b010c1872a
Partially-Implements: blueprint scaling-group-common