Use sections for individual properties and attributes
It is often helpful to point people at the documentation
for a specific attribute or property. This change achieves
this by using sections instead if definition items to
format the documentation of a property/attribute.

Complex property schemas are indented using block quotes
and the fragment identifier is appended to so that each
link is unique. (eg, #OS::Ceilometer::Alarm-prop-query-*-field)

The 'required' and 'optional' sections have been brought up one
level, replacing 'Properties' with 'Required Properties' and
'Optional Properties'.

Closes-Bug: #1486429
Change-Id: Ifd497aac74e8aff76ea91a295df17bb1da949ba8