Raise exception if asg in cooldown
The signal() method of resource class doesn't add
event to event list if the signal can not be performed i.e.
handle_signal raise exception. So it is needed to raise exception
if asg in cooldown. Otherwise the event will be added to event list
in SIGNAL_COMPLETE state, but no action will be performed.

Change-Id: Iebdb6c9231aef90d11428ca6beed150cfba2715f
Closes-Bug: #1486945