Don't metadata_update all resources for deployment signals
This fixes a race when a call to
SoftwareConfigService._push_metadata_software_deployments is in
progress at the same time as other deployments are signalling heat.

The signal triggers a metadata_update() on the server resource
which results in rsrc_metadata being updated without the resource
lock. This can overwrite the rsrc_metadata being written by
SoftwareConfigService._push_metadata_software_deployments resulting in
stale metadata.

Change-Id: I081bc154ed7e79f4a4258c846857b3f39cc7887c
Closes-Bug: #1488366