Split out calculation of the next batch into a separate function
Create a next_batch function in the scaling library that calculates only
what the next batch update should look like. For now, continue to call this
from the get_batches function so that we can verify there is no change in
behaviour. However, this will make it easier to add improvements in the
future, and to reconcile InstanceGroup with ResourceGroup.

Change-Id: I0fe25556de2a678ead14bc64f77c0466fe8f9e92
Partially-Implements: blueprint scaling-group-common