Don't repeatedly overwrite definitions in RG rolling update
On each batch, we were returning the names of all the resources that had
been added/updated on previous batches, as well as the ones to add/update
in this batch. This patch changes it so we only return the names of the
resources to be added/updated in this batch. This should result in no
change in behaviour, since the previously-updated resources will retain
their existing (updated) definition.

Change-Id: I4879303ec0ae74bfe929429fd8844bb37d30cf13
Partially-Implements: blueprint scaling-group-common