Use the member_definition library func in ResourceGroup rolling updates
Use the scaling library function for generating the intermediate template
stages during a rolling update, instead of maintaining a separate
implementation for ResourceGroup.

One improvement here is that if we are scaling down at the same time as
doing a rolling update and we already have more members than we need to
maintain the minimum in service, then we'll scale down to the minimum we
need immediately.

Change-Id: Ied7abb0a8d9ccfbda66d2dbee75057643ba7a169
Partially-Implements: blueprint scaling-group-common