Allow ResourceGroup rolling update to scale down
The rolling update code is capable of scaling the ResourceGroup down to the
correct size at the end of the rolling update, so allow it do so rather
than appending a separate scale-down step.

This change makes the _get_batches() function as simple as it can be - it
now differs from the one used by InstanceGroup only in that resources that
will later be deleted (because their indices are too high) are not counted
as updated.

Change-Id: I00649eaa685cfa7d05a8471ec5e5d287ac046357
Partially-Implements: blueprint scaling-group-common