Merge Server os-collect-config metadata on update
Currently if someone does this, then updates the stack,
we overwrite *all* the server metadata, which will be bad for
all user_data_format: SOFTWARE_CONFIG use-cases, because we want
the transport details to remain over updates, and for the
deployments list to be untouched.

  aserver:
    type: OS::Nova::Server
    metadata:
      foo: 123

Note that we make an effort to merge the os-collect-config data
because it may be valid to pass additional occ options via the
template, in addition to those created internally.

Currently we silently overwrite the "deployments" key, which is
consistent with what we do on create, but a validation check
would be a good addition to fail-fast if we're about to stomp
on a template provided key.

Change-Id: Icf6001dde0a7633f85f5c306e31454d6457402da
Partial-Bug: #1488596