Allow map/list items for list_join
Currently any attempt to join json/map or list parameters/attributes
results in an error, which is inconvenient if you wish to join
several paramters/attributes into a single string that may be easily
consumed in an instance (e.g by splitting each element on a delimiter
then using jq or python to process the json).

So, tolerate non-string list items, and attempt to serialize them
as json, then join the resulting json strings.

Partial-Bug: #1489028
Change-Id: I699fa2ec43d173fff1f06887aaa80e3fbac8d668