Assign parent_stack during validation
This patch is to fix call resource_facade failed.
The root cause is that call Stack.parent_resource failed,
because nest stack's _parent_stack is None in stack validation phase.

The call stack is as follow:
---------------------------------------------------
 EngineService.create_stack
 ->EngineService._parse_template_and_validate_stack
   ->Stack.validate
     ->TemplateResource.validate
       ->StackResource.validate
         ->StackResource.validate_nested_stack
           ->StackResource._parse_nested_stack
             ->parser.Stack
---------------------------------------------------
a) Here a nest stack is created, but the own_id is None.
b) Then when continue the nest stack, it will finally
call Stack.parent_resource, which will raise an exception
because the nest stack's _parent_stack member is None.

Change-Id: I0a886265a22a2a93e2bf286005e0e7ef7a6f498f
Closes-Bug: #1491146