Only write to template files if template data changes
Currently the template data is assigned to the files dict regardless
of whether the data has changed or not.

This has no consequence on Liberty, but the files dict in Kilo
and Juno is a sqlalchemy MutableDict so these assignments are triggering
updates to the raw_template table. These updates are even triggered for
calls to heat resource-list. With large raw_template records this can make
heat unusable.

Therefore this fix is presented here as having no effect (and being
harmless) to git master, but high priority backports for Kilo and Juno.

Change-Id: Ibb714a9c5dc9f6879ff108419127da985a57897b
Closes-Bug: #1494108
(cherry-picked from 779ba8e431aecac3c451c08aedcc0ad8cf0a4449)