Correctly filter events on resource name
Filtering on resource name can be accomplished by using the db
filters instead of using a filter method. Furthermore, using a
custom filter breaks certain assumptions made by other parts of
the codebase, which lead to pagination being broken.
Specifically, custom filter for resource name is applied only
after other filtering is finished. As a result, we try to find
result from only 1 event and this event belongs to whole
stack and thus its resource_name is equal to the stack name.
This returns an event not found error instead of returning
the first event assosciated with the resource name.

Change-Id: Ie7d28ffb88a79d6e0f516f0db385be2eed59f7f0
Closes-Bug: #1502751