Do no net/subnet checks during validate
Currently validation fails if the network or the subnet are defined in
the same template as the server because they don't exist during
validate.

_validate_belonging_subnet_to_net is called during _build_nics so this
check still happens when the required resources actually exist.

Also, test method validate_internal_port_subnet_not_this_network was not
running because its name was not prefixed with test_. Fixing and
updating this test revealed issues with the validation failed message.

Closes-Bug: #1503060
Change-Id: I2e062abf5b9fb7446b979ba6acc2e2b531d957d6