Fix parameter_defaults for comma_delimited_list and json
When environment of a stack contained a parameter_defaults section with
a value for a parameter of type comma_delimited_list or json, that default
didn't get applied unless in a nested stack.

The issue was that CommaDelimitedListParam contains a variable to cache
its value, and that cache was only created upon creation of a
CommaDelimitedListParam object, but it wasn't updated on `set_default`
method call. The cache now updates on `set_default` too.

Co-Authored-By: Steven Hardy <shardy@redhat.com>
Change-Id: I85b13a79dbc97a77e20c0d5df8eaf05b3000815e
Closes-Bug: #1504137
(cherry picked from commit 3827c61b25fa86c1f9882be15caf27d2efa6c4fe)