Fix race condition in stack listing
When listing stacks, we first load them from the database, and then we
load their template in a different transaction. If a stack is deleted in
the mean time, loading the template will raise a NotFound exception, and
will make the listing returns NotFound, instead of ignoring that stack.
This patch handles the exception and continue to try to load stacks.

Change-Id: Iba6db81e3c10310955928fe723b84b5f62fc3fc1
Closes-Bug: #1508390