Resource.has_interface check get_resource_info result
This changes the behaviour of has_interface to not assume that
get_resource_info always returns a type. There are circumstances when it
won't, such as when the type is no longer in the new registry during a
stack update.

Change-Id: Idd9dade85d105dc6839ff0740e6c80aca6b7ad9f
Closes-Bug: #1509880
(cherry picked from commit af3355d34c8b8e4030f7a928a6c7cb9ad70e0795)