Reset stack status even when lock engine_id is None
Having a stack which is wedged UPDATE_IN_PROGRESS due to another bug,
restarting heat-engine should have resulted in the stack being reset to
UPDATE_FAILED. However the reset was skipped because the lock engine_id
was None.

This change removes the engine_id check since it is possible in practice
for an UPDATE_IN_PROGRESS stack to not be locked.

Since it is possible for the stack to change state in the unlocked
window, a check is added that the stack is still IN_PROGRESS before the
reset is performed.

Change-Id: I1739ccbdf75af35aac5be16b99200975df58b8e2
Closes-Bug: #1514615
(cherry picked from commit 44b2f0965dd329cf45a151705b0648819acff8c7)