Copy the env to the backup stack in failed update
When an update fails, we currently update the environment of the stack
to contain both the new and old parameters and types. Unfortunately we
don't do that for the backup stack which is reused. This patch adds the
environment change.

Change-Id: I4dc5dd35e4aeee498dd8960a1000913e97b924d5
Closes-Bug: 1508096
(cherry picked from commit 88da460316e2c18b3ceb58f51dcf0bda717a18a3
                       and 13c68910704844719b2d3de353d2db276377bf11)