Fix stack lock ActionInProgress in the admin case
When the engine starts and tries to reset the status of stacks in
progress, it tries to acquire the stack lock on them. The stack lock
queries the database to retrieve the name of the stack, but does a
tenant-scope query, thus fails when as we use the admin context during
startup. This fixes it by passing tenant_safe=False.

Change-Id: I4393fdb601eedb36d7288afd57fb5a1c96af7216
Closes-Bug: #1516712
(cherry picked from commit 9924d780e82d8419783177094d63c8333482a129)