Revert "Fix wrong type of exception raised"
This reverts commit c5bb4d2ab83820f5ee2ac3fc6e8d9210e98b26ef.

Every part of this patch was wrong, except that it did (by accident)
start returning 500 errors instead of 400 errors in the case where a
template in the global environment was not found, which is actually the
operator's fault and not the user's. That's not really a serious enough
problem to justify changing it on a stable branch though, so this is a
straight revert. Luckily the patch hasn't made it into a release yet.

Change-Id: Ia3154c7c41bb69cfa2ee36eb6d48fcd339900d60
Closes-Bug: #1518458
Related-Bug: #1447194