Add a separate get_class_to_instantiate() method to Environment
We were previously using get_class for two different purposes - to get a
resource plugin on which we could perform introspection to obtain the
properties and attributes schema, and to get a resource plugin we could
instantiate to create a Resource object. These are both the same except in
the case of a TemplateResource, where having two different use cases for
the same piece of code was adding considerable extra complexity. Combining
the use cases in this way also made the error handling confusing (leading
to bug 1518458).

This change separates out the two cases.

Change-Id: I3bde081cd4537810c8c6a0948ab447c3fb7ca9bc
Related-Bug: #1447194
Related-Bug: #1518458
Related-Bug: #1508115