Make Server software_config_transport updatable
This change makes it possible to update a server's
software_config_transport property without replacing the server.

Some reasons for wanting to do this with long-running servers include:

- moving to POLL_TEMP_URL to reduce load on the heat engines, or to be
  able to create deployments outside the stack
- moving to ZAQAR_MESSAGE once the host cloud has deployed zaqar

Any existing transport *must* continue working and return up-to-date
data for the lifecycle of the stack since it can't be known when all
servers have switched to the new transport. This means that a transport's
enabled status should be determined by the enabling resource data
rather than the software_config_transport property.

The os-collect-config configuration data needs to ensure that old
transport configurations are disabled. This includes local boot config
which remains on the server for its lifetime, and is outdated once the
transport changes. This disabling is achieved by setting any existing
configuration values to None.

This has been tested on a tripleo cloud by repeatedly switching between
POLL_TEMP_URL and POLL_SERVER_CFN during a stack update and confirming
that os-collect-config switches to the new transport.

Change-Id: I9c475f0c489c67db5895924050186228403f2773
Closes-Bug: #1519609