Immutable parameters
Add a new "immutable" boolean field to the parameters section in a HOT
template.  A value of True will result in the engine rejecting
stack-updates that include changes to that parameter.  When not
specified in the template, "immutable" defaults to False to ensure
backwards compatibility with old templates.

blueprint: immutable-parameters
Change-Id: If17df0b97e96ff2a1926451634e4744d6c3e26f1