Explicitly naming resources in "unsupported" warnings
Warnings that a resource is not supported are now prefaced
with the resource name and status (e.g., DEPRECATED). In
addition, the warning is now logged via LOG.warn(_LW()).

Previously, warnings.warn(six.text_type()) was used. In
addition to being routed to the log file, it included a
confusing reference to the last line of invocation, which
was the fragment "info.value.support_status.message))".

Change-Id: I006120a395007d88f3ba04a4af89283a1a077c3f
Closes-Bug: #1524512