Fix property comparison in neutron resources
We were trying to validate deprecated neutron properties with
properties.data, which work when you have a reference, but not a
parameter. Instead, compare it None which is what's the value when it's
not set at all.

We can't backport the original patch because it uses the new translation
mechanism for deprecating properties.

Change-Id: Ifa96a1cf94ced833afaf4b7ca306a953987548b6
Closes-Bug: #1518676