Make metadata_server_url config optional
Similar to If8a2d3f37d87c26228e709c20f61969b397f2da0 which made
the heat_waitcondition_server_url optional, this makes it possible
to not set the heat_waitcondition_server_url option in heat.conf,
and instead have heat derive it from the CFN endpoint in the
service catalog.

Arguably this is a more correct approach in the majority of cases, but
the config file overrides are maintained to enable environments which
require them to continue using them (such as where instances must
route requests to heat via a specific proxy not defined in the service
catalog).

Change-Id: Id402664e38e3da071ad634233b3a1f8e13af152d