InstanceGroup to keep FAILED resources in template
Modify the InstanceGroup (and, by extension, Heat and AWS
AutoscalingGroup) types to give members in a FAILED state the highest
priority for being removed when scaling down or being updated in
a rolling update. Currently, FAILED resources are omitted when building
a new template for the scaling group, so any such resources would never
be replaced by one of the same name. This change will allow for
continuity of naming in the case of a change that doesn't permanently
remove the resource due to scaling down.

Change-Id: I8fe93869dfbe4c4de6c48a65aeef0f71355eaf46
Partially-implements: blueprint mark-unhealthy