Updated validation for OS::Trove::Instance
Trove supports using a default datastore version if none is specified in
the request.  The OS::Trove::Instance resource was raising an exception
rather than allowing Trove to use the default version in the event that
there were multiple versions of a specific datastore available and a
version was not specified.

By removing the section of code that performed this validation, we can
allow Trove to determine which datastore version to use.

Change-Id: I0ff1ebd1e7d70ea0a98630584174cf8ce1ce6abd
Closes-Bug: #1542437