Ignore errors on old properties during update
When an update fails on a template resource, we still store the template
in the stack environment. Retrying the update will fail because of
missing parameters, but we can ignore those as they should be harmless
and only matter in this broken case.

Change-Id: I859ea72b1cc95162b6f498b6af65e4d9fdd9458f
Closes-Bug: #1543685
(cherry picked from commit 1cfde62b2ff8a60c8cc3a9ef4926e5f4a7116aee)